refactor(commands): Move configuration migration to a dedicated command
In preparation for adding more options to migrate ORT configuration,
split the existing migration option out of the `config` tool to a
dedicated `migrate` tool. The `config` tool now remains "read-only".
