+++ plugins/commands/migrate/src/main/resources/META-INF/services/org.ossreviewtoolkit.plugins.commands.api.OrtCommand
@@ -1,0 +1,1 @@
+org.ossreviewtoolkit.plugins.commands.migrate.MigrateCommand
